requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.3.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
tqdm>=4.64.0
psutil>=5.9.0 